+++ src/main/java/com/main/cloudapi/controller/ComplectationController.java
@@ -1,16 +1,19 @@
 package com.main.cloudapi.controller;
 
 import com.main.cloudapi.api.ComplectationControllerI;
+import com.main.cloudapi.constmes.ThrowFabric;
 import com.main.cloudapi.controller.base.BaseController;
 import com.main.cloudapi.entity.Brand;
 import com.main.cloudapi.entity.Complectation;
 import com.main.cloudapi.entity.views.ComplectationView;
 import com.main.cloudapi.service.ComplectationService;
+import com.main.cloudapi.service.UserService;
 import com.main.cloudapi.utils.JsonWebContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
 
@@ -22,6 +25,9 @@ public class ComplectationController extends BaseController implements Complecta
 
     @Autowired
     ComplectationService complectationService;
+
+    @Autowired
+    UserService userService;
 
     @Override
     public Complectation getComplectation(@PathVariable String brandID, @PathVariable String carID, @PathVariable String id) {
@@ -36,17 +42,32 @@ public class ComplectationController extends BaseController implements Complecta
     }
 
     @Override
-    public Complectation addComplectation(@PathVariable String brandID, @PathVariable String carID, @RequestBody String json) {
+    public Complectation addComplectation(@PathVariable String brandID, @PathVariable String carID, @RequestBody String json,
+                                          @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return complectationService.addComplectation(parseID(brandID), parseID(carID), json);
     }
 
     @Override
-    public Complectation editComplectation(@PathVariable String brandID, @PathVariable String carID, @PathVariable String id, @RequestBody String json) {
+    public Complectation editComplectation(@PathVariable String brandID, @PathVariable String carID, @PathVariable String id, @RequestBody String json,
+                                           @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return complectationService.editComplectation(parseID(brandID), parseID(carID), parseID(id), json);
     }
 
     @Override
-    public Complectation deleteComplectation(@PathVariable String brandID, @PathVariable String carID, @PathVariable String id) {
+    public Complectation deleteComplectation(@PathVariable String brandID, @PathVariable String carID, @PathVariable String id,
+                                             @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return complectationService.deleteComplectation(parseID(brandID), parseID(carID), parseID(id));
     }
 }
