+++ src/main/java/com/main/cloudapi/controller/PriceCoefController.java
@@ -1,13 +1,16 @@
 package com.main.cloudapi.controller;
 
 import com.main.cloudapi.api.PriceCoefControllerI;
+import com.main.cloudapi.constmes.ThrowFabric;
 import com.main.cloudapi.controller.base.BaseController;
 import com.main.cloudapi.entity.PriceCoef;
 import com.main.cloudapi.service.PriceCoefService;
+import com.main.cloudapi.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
 
@@ -20,6 +23,9 @@ public class PriceCoefController extends BaseController implements PriceCoefCont
     @Autowired
     PriceCoefService priceCoefService;
 
+    @Autowired
+    UserService userService;
+
     @Override
     public PriceCoef getPriceCoef(@PathVariable String id) {
         return priceCoefService.getById(parseID(id));
@@ -31,17 +37,32 @@ public class PriceCoefController extends BaseController implements PriceCoefCont
     }
 
     @Override
-    public PriceCoef addPriceCoef(@RequestBody String json) {
+    public PriceCoef addPriceCoef(@RequestBody String json,
+                                  @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return priceCoefService.add(json);
     }
 
     @Override
-    public PriceCoef editPriceCoef(@PathVariable String id, @RequestBody String json) {
+    public PriceCoef editPriceCoef(@PathVariable String id, @RequestBody String json,
+                                   @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return priceCoefService.edit(parseID(id), json);
     }
 
     @Override
-    public PriceCoef deletePriceCoef(@PathVariable String id) {
+    public PriceCoef deletePriceCoef(@PathVariable String id,
+                                     @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return priceCoefService.delete(parseID(id));
     }
 }
