+++ src/main/java/com/main/cloudapi/controller/SettingsController.java
@@ -1,13 +1,16 @@
 package com.main.cloudapi.controller;
 
 import com.main.cloudapi.api.SettingsControllerI;
+import com.main.cloudapi.constmes.ThrowFabric;
 import com.main.cloudapi.controller.base.BaseController;
 import com.main.cloudapi.entity.Settings;
 import com.main.cloudapi.service.SettingsService;
+import com.main.cloudapi.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
 
@@ -20,6 +23,9 @@ public class SettingsController extends BaseController implements SettingsContro
     @Autowired
     SettingsService settingsService;
 
+    @Autowired
+    UserService userService;
+
     @Override
     public Settings getSetting(@PathVariable String id) {
         return settingsService.getById(parseID(id));
@@ -31,17 +37,32 @@ public class SettingsController extends BaseController implements SettingsContro
     }
 
     @Override
-    public Settings addSetting(@RequestBody String json) {
+    public Settings addSetting(@RequestBody String json,
+                               @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return settingsService.addSetting(json);
     }
 
     @Override
-    public Settings editSetting(@PathVariable String id, @RequestBody String json) {
+    public Settings editSetting(@PathVariable String id, @RequestBody String json,
+                                @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return settingsService.editSetting(parseID(id), json);
     }
 
     @Override
-    public Settings deleteSetting(@PathVariable String id) {
+    public Settings deleteSetting(@PathVariable String id,
+                                  @RequestParam(value = "access_token", defaultValue = "", required = false)String access_token) {
+        userService.validateToken(access_token);
+        if (!userService.checkPermission(userService.getCurUser().getId())){
+            throw new ThrowFabric.LockedException("Permission denied");
+        }
         return settingsService.deleteSetting(parseID(id));
     }
 }
